Twitch API Rework
