+++ commands/other/twitchstatus.js
@@ -2,11 +2,8 @@ const fetch = require('node-fetch');
 const { Command } = require('discord.js-commando');
 const { twitchClientID, twitchClientSecret } = require('../../config.json');
 const { MessageEmbed } = require('discord.js');
-
-if (twitchClientID == null || twitchClientSecret == null)
-  return console.log(
-    'INFO: TwitchStatus command removed from the list. \nMake sure you have twitchClientID and twitchToken in your config.json to use TwitchStatus command!'
-  );
+const TwitchAPI = require('../../resources/twitch/twitch-api.js');
+if (twitchClientID == null || twitchClientSecret == null) return;
 module.exports = class TwitchStatusCommand extends Command {
   constructor(client) {
     super(client, {
@@ -35,7 +32,7 @@ module.exports = class TwitchStatusCommand extends Command {
     const textFiltered = textRaw.replace(/https\:\/\/twitch.tv\//g, '');
     let access_token;
     try {
-      access_token = await TwitchStatusCommand.getToken(
+      access_token = await TwitchAPI.getToken(
         twitchClientID,
         twitchClientSecret,
         scope
@@ -46,7 +43,7 @@ module.exports = class TwitchStatusCommand extends Command {
     }
 
     try {
-      var user = await TwitchStatusCommand.getUserInfo(
+      var user = await TwitchAPI.getUserInfo(
         access_token,
         twitchClientID,
         textFiltered
@@ -58,7 +55,7 @@ module.exports = class TwitchStatusCommand extends Command {
 
     const user_id = user.data[0].id;
     try {
-      var streamInfo = await TwitchStatusCommand.getStream(
+      var streamInfo = await TwitchAPI.getStream(
         access_token,
         twitchClientID,
         user_id
@@ -103,11 +100,24 @@ module.exports = class TwitchStatusCommand extends Command {
       return;
     }
     try {
-      var gameInfo = await TwitchStatusCommand.getGames(
+      var gameInfo = await TwitchAPI.getGames(
         access_token,
         twitchClientID,
         streamInfo.data[0].game_id
       );
+      var result = await probe(
+        gameInfo.data[0].box_art_url.replace(/-{width}x{height}/g, '')
+      );
+      var canvas = Canvas.createCanvas(result.width, result.height);
+      var ctx = canvas.getContext('2d');
+      // Since the image takes time to load, you should await it
+      var background = await Canvas.loadImage(
+        gameInfo.data[0].box_art_url.replace(/-{width}x{height}/g, '')
+      );
+      // This uses the canvas dimensions to stretch the image onto the entire canvas
+      ctx.drawImage(background, 0, 0, canvas.width, canvas.height);
+      // Use helpful Attachment class structure to process the file for you
+      var attachment = new MessageAttachment(canvas.toBuffer(), 'box_art.png');
     } catch (e) {
       message.say(e);
       return;
@@ -134,13 +144,9 @@ module.exports = class TwitchStatusCommand extends Command {
           .replace(/{width}x{height}/g, '1280x720')
           .concat('?r=' + Math.floor(Math.random() * 10000 + 1))
       )
-      .setTimestamp(streamInfo.data[0].started_at);
-    if (gameInfo.data[0].box_art_url.search(/.jpg/g))
-      onlineEmbed.setThumbnail(user.data[0].profile_image_url);
-    else
-      onlineEmbed.setThumbnail(
-        gameInfo.data[0].box_art_url.replace(/-{width}x{height}/g, '')
-      );
+      .setTimestamp(streamInfo.data[0].started_at)
+      .attachFiles(attachment)
+      .setThumbnail('attachment://box_art.png');
     if (user.data[0].broadcaster_type == '')
       onlineEmbed.addField('Rank:', 'BASE!', true);
     else {
@@ -152,112 +158,5 @@ module.exports = class TwitchStatusCommand extends Command {
     }
     message.say(onlineEmbed);
     return;
-  }
-  static getToken(twitchClientID, twitchClientSecret, scope) {
-    return new Promise(async function fetchToken(resolve, reject) {
-      try {
-        const response = await fetch(
-          `https://id.twitch.tv/oauth2/token?client_id=${twitchClientID}&client_secret=${twitchClientSecret}&grant_type=client_credentials&scope=${scope}`,
-          {
-            method: 'POST'
-          }
-        );
-        const json = await response.json();
-        if (json.status == 400) {
-          reject(
-            'Something went wrong when trying to fetch a twitch access token'
-          );
-        } else {
-          resolve(json.access_token);
-        }
-      } catch (e) {
-        console.error(e);
-        reject('There was a problem fetching a token from the Twitch API');
-      }
-    });
-  }
-  static getUserInfo(token, client_id, username) {
-    return new Promise(async function fetchUserInfo(resolve, reject) {
-      try {
-        const response = await fetch(
-          `https://api.twitch.tv/helix/users?login=${username}`,
-          {
-            method: 'GET',
-            headers: {
-              'client-id': `${client_id}`,
-              Authorization: `Bearer ${token}`
-            }
-          }
-        );
-        const json = await response.json();
-        if (json.status == `400`) {
-          reject(`:x: ${username} was Invaild, Please try again.`);
-          return;
-        }
-
-        if (json.status == `429`) {
-          reject(`:x: Rate Limit exceeded. Please try again in a few minutes.`);
-          return;
-        }
-
-        if (json.status == `503`) {
-          reject(
-            `:x: Twitch service's are currently unavailable. Please try again later.`
-          );
-          return;
-        }
-
-        if (json.data[0] == null) {
-          reject(`:x: Streamer ${username} was not found, Please try again.`);
-          return;
-        }
-        resolve(json);
-      } catch (e) {
-        console.error(e);
-        reject('There was a problem fetching user info from the Twitch API');
-      }
-    });
-  }
-  static getStream(token, client_id, userID) {
-    return new Promise(async function fetchStream(resolve, reject) {
-      try {
-        const response = await fetch(
-          `https://api.twitch.tv/helix/streams?user_id=${userID}`,
-          {
-            method: 'GET',
-            headers: {
-              'client-id': `${client_id}`,
-              Authorization: `Bearer ${token}`
-            }
-          }
-        );
-        const json = await response.json();
-        resolve(json);
-      } catch (e) {
-        console.error(e);
-        reject('There was a problem fetching stream info from the Twitch API');
-      }
-    });
-  }
-  static getGames(token, client_id, game_id) {
-    return new Promise(async function fetchGames(resolve, reject) {
-      try {
-        const response = await fetch(
-          `https://api.twitch.tv/helix/games?id=${game_id}`,
-          {
-            method: 'GET',
-            headers: {
-              'client-id': `${client_id}`,
-              Authorization: `Bearer ${token}`
-            }
-          }
-        );
-        const json = await response.json();
-        resolve(json);
-      } catch (e) {
-        console.error(e);
-        reject('There was a problem fetching stream info from the Twitch API');
-      }
-    });
   }
 };
